+++ src/analytics/index.js
@@ -1,6 +1,6 @@
 import "../styles/analytics.css";
 
-import { Statistics } from "../js/components/analytics/Statistics.js";
+import { Statistics } from "../blocks/analytics/Statistics.js";
 import { DataStorage } from "../js/modules/DataStorage.js";
 import { DAYS_WEEK_NOW } from "../js/constants/dates.js";
 
@@ -9,6 +9,9 @@ const dataStorage = new DataStorage;
 const statistics = new Statistics(dataStorage.getNewsDataObj(), localStorage['NewsAnalyzer reqPhrase'], localStorage['NewsAnalyzer newsTotalResults']);
 
 //Рендер статистики и заголовков при загрузке страницы
-statistics.renderStatTitles();
-statistics.renderStatDates(DAYS_WEEK_NOW());
-statistics.renderStatValues();
+window.onload = () => {
+  statistics.renderStatTitles();
+  statistics.renderStatDates(DAYS_WEEK_NOW());
+  statistics.renderStatValues();
+}
+
